Reverted utils and only extracted required function
