+++ app/react/Markdown/components/Connect.js
@@ -2,7 +2,7 @@ import { connectAdvanced } from 'react-redux';
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import PagesContext from './Context';
-import { objectPath } from '../utils.tsx';
+import { objectPath } from '../utils';
 
 class Connect extends Component {
   static objectPath(path, object) {
